requirements.txt
agno
openai
google-generativeai
google-genai
fastapi
uvicorn
prisma
python-dotenv
pydantic

# NEW: Authentication & Email
clerk-backend-api>=0.2.0
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.0
python-multipart>=0.0.6
aiosmtplib>=3.0.0
email-validator>=2.1.0